특정 디렉토리에 이미 동일한 이름의 파일이 존재하면 복사를 진행하지 않는 코드 작성

+++ FileOI.py
@@ -13,8 +13,22 @@ def get_all_file_path(input_dir, file_extension):
     return temp
 
 
+def get_pure_filename(filename):
+    temp = filename.split('.')
+    del temp[-1]
+    temp = ' '.join(temp)
+    temp = temp.split('/')
+    temp = temp[-1]
+    return temp
+
+
 def copy_all_file_to_target_directory(data, target_directory):
-    shutil.copy(data, target_directory)
+    pure_filename = get_pure_filename(data)
+    temp = '{}/{}.txt'.format(target_directory, pure_filename)
+    if os.path.isfile(temp):
+        pass
+    else:
+        shutil.copy(data, target_directory)
     return 0
 
 
@@ -23,7 +37,12 @@ def move_all_file_to_target_directory(data, target_directory):
     return 0
 
 
-def parallel_preprocess(filelist, target_directory, type='copy', parallel=None):
+def remove_all_file_to_target_directory(data, target_directory):
+    os.remove(data)
+    return 0
+
+
+def parallel_preprocess(filelist, target_directory, _type='copy', parallel=None):
 
     try:
         if not (os.path.isdir(target_directory)):
@@ -33,9 +52,30 @@ def parallel_preprocess(filelist, target_directory, type='copy', parallel=None):
             print("Failed to create directory!!!!!")
             raise
     with multiprocessing.Pool(parallel) as p:
-        if type == 'copy':
+        if _type == 'copy':
             func = functools.partial(copy_all_file_to_target_directory, target_directory=target_directory)
             output = list(tqdm(p.imap(func, filelist), total=len(filelist)))
-        elif type == 'move':
+        elif _type == 'move':
             func = functools.partial(move_all_file_to_target_directory, target_directory=target_directory)
             output = list(tqdm(p.imap(func, filelist), total=len(filelist)))
+        elif _type == 'remove':
+            func = functools.partial(remove_all_file_to_target_directory, target_directory=target_directory)
+            output = list(tqdm(p.imap(func, filelist), total=len(filelist)))
+
+
+def wav_text_pair(input_dir):
+    wav = get_all_file_path(input_dir, file_extension='wav')
+    txt = get_all_file_path(input_dir, file_extension='txt')
+    # for file in tqdm(wav):
+    #     temp = file.replace('.wav', '.txt')
+    #     if temp not in txt:
+    #         print(file, "not exist")
+    for idx, file in enumerate(wav):
+        wav[idx] = file.replace('.wav', '')
+    for idx, file in enumerate(txt):
+        txt[idx] = file.replace('.txt', '')
+
+    wav = set(wav)
+    txt = set(txt)
+    print(wav.difference(txt))
+    print(txt.difference(wav))
